[refactor] mock 코드 이동
